Change DeleteLog func to take in int64 for log id

+++ internal/database/database.go
@@ -50,7 +50,7 @@ type UserLogDatabase interface {
 	//
 	// Returns true of the operation was successful.
 	// Else, it returns false and an error if it was unsuccessful.
-	DeleteLog(int, int) (bool, error)
+	DeleteLog(int, int64) (bool, error)
 }
 
 type UserDatabase interface {
@@ -246,7 +246,7 @@ func (db Sqlite3DB) UpdateLog(newLog model.Log) (bool, error) {
 	return true, nil
 }
 
-func (db Sqlite3DB) DeleteLog(userId int, id int) (bool, error) {
+func (db Sqlite3DB) DeleteLog(userId int, id int64) (bool, error) {
 	statement, err := db.Prepare("DELETE FROM 'logs' WHERE id = ? AND userId = ?")
 
 	if err != nil {
